server/requirements.txt
django==5.2.11
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
scikit-learn
numpy
openai
